# api/openapi.yaml
openapi: 3.1.0
info:
  title: Inventory Manager
  description: >
    Inventory Manager that stores items, manipulates their quantities,
    and calculates the value of inventory items.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local test server
paths:
  /items:
    get:
      description: Fetches all items in the inventory.
      responses:
        '200':
          description: Returns the items as a JSON object keyed by item ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Items'
        '500':
          description: An unexpected server error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /items/value:
    get:
      description: Calculates the total value of items in the inventory.
      parameters:
        - name: ids
          in: query
          description: >
            The IDs of the items to calculate the value for. 
            If omitted, the value of all items in the inventory will be calculated.
          required: false
          schema:
            type: array
            items:
              type: integer
              format: int64
              minimum: 0
          style: form
      responses:
        '200':
          description: Returns the value as a JSON object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Value'
        '500':
          description: An unexpected server error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /item:
    post:
      description: Adds the specified item to inventory.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Item"
      responses:
        '201':
          description: The item was added successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultItem'
        '400':
          description: The given data was invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: An unexpected server error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /item/{id}:
    patch:
      description: Alters the quantity associated to an Item in the inventory.
      parameters:
        - name: id
          in: path
          description: The ID of the item to manipulate.
          required: true
          schema:
            example: 1
            type: integer
            format: int64
            minimum: 0
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Delta"
      responses:
        '200':
          description: Returns the updated Item on success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultItem'
        '400':
          description: The given id is not an unsigned int.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: The given id was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: The given delta would cause the quantity to become negative.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      description: Deletes an item from the inventory.
      parameters:
        - name: id
          in: path
          description: The ID of the item to delete.
          required: true
          schema:
            example: 1
            type: integer
            format: int64
            minimum: 0
      responses:
        '204':
          description: The item was deleted successfully.
        '400':
          description: The given id is not an unsigned int.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: The given id was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: An unexpected server error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Items:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/Item'
      example:
        1:
          name: 'Table'
          quantity: 10
          price_cents: 11999
        2:
          name: 'Chair'
          quantity: 5
          price_cents: 2999

    Item:
      type: object
      properties:
        name:
          type: string
          example: 'Table'
          description: The name of the inventory item.
        quantity:
          type: integer
          format: int64
          example: 10
          description: The current quantity of the inventory item.
        price_cents:
          type: integer
          format: int64
          example: 11999
          description: The monetary value of the inventory item in euro.
      required: [name, quantity, price_cents]

    ResultItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
          minimum: 0
          example: 1
          description: The ID of the inventory item.
        name:
          type: string
          example: 'Table'
          description: The name of the inventory item.
        quantity:
          type: integer
          format: int64
          example: 10
          description: The current quantity of the inventory item.
        price_cents:
          type: integer
          format: int64
          example: 11999
          description: The monetary value of the inventory item in euro.

    Delta:
      type: object
      properties:
        quantity_delta:
          type: integer
          format: int64
          example: -5
      required: [quantity_delta]

    Value:
      type: object
      properties:
        success:
          type: boolean
          example: true
        value:
          type: integer
          format: int64
          example: 124890

    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: 'Error message.'